rename cmds
