--- cmd/client_test/html/login.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html>
-  <form action="/login" method="post">
-    <input
-      name="auth-input"
-      id="auth-input"
-      placeholder="Enter your handle, did, or pds url"
-    />
-    <button type="submit" value="Submit">Submit</button>
-  </form>
-</html>
