--- cmd/client_test/html/make-post.html
@@ -1,5 +1,0 @@
-<!doctype html>
-<html>
-  <p>Successfully created a post!</p>
-  <p><a href="/">Go to home</a></p>
-</html>
