--- cmd/client_test/html/profile.html
@@ -1,7 +1,0 @@
-<!doctype html>
-<html>
-  <h2>{{ .DisplayName }}</h2>
-  <p>{{ .Handle }}</p>
-  <p>{{ .Description }}</p>
-  <a href="/">Go Home</a>
-</html>
