+++ cmd/web_server_demo/html/login.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html>
+  <form action="/login" method="post">
+    <input
+      name="auth-input"
+      id="auth-input"
+      placeholder="Enter your handle, did, or pds url"
+    />
+    <button type="submit" value="Submit">Submit</button>
+  </form>
+</html>
