+++ cmd/web_server_demo/html/make-post.html
@@ -1,0 +1,5 @@
+<!doctype html>
+<html>
+  <p>Successfully created a post!</p>
+  <p><a href="/">Go to home</a></p>
+</html>
