+++ cmd/web_server_demo/html/profile.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+  <h2>{{ .DisplayName }}</h2>
+  <p>{{ .Handle }}</p>
+  <p>{{ .Description }}</p>
+  <a href="/">Go Home</a>
+</html>
